+++ packages/qdadm/src/chain/index.js
@@ -1,0 +1,12 @@
+/**
+ * Chain Module - Active navigation stack management
+ *
+ * Provides:
+ * - ActiveStack: Reactive container for current navigation stack
+ * - useActiveStack: Composable to build and access the stack
+ *
+ * @module chain
+ */
+
+export { ActiveStack } from './ActiveStack.js'
+export { useActiveStack } from './useActiveStack.js'